+++ agaein_web/codegen.yml
@@ -1,6 +1,6 @@
 overwrite: true
-# schema: 'https://www.agaein.com/graphql'
-schema: 'http://localhost:3005/graphql'
+schema: 'https://www.agaein.com/graphql'
+# schema: 'http://localhost:3005/graphql'
 documents:
     - 'src/graphql/queries/**/*.graphql'
 generates:
