--- agaein_web/src/components/pages/article/CreateArticle/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CreateArticle';
