+++ agaein_web/src/router/index.tsx
@@ -12,7 +12,9 @@ function Router() {
             <Route path="/" component={Home} exact />
             <Route path="/articles/" component={ArticleList} />
             <Route path="/createArticle/step1" component={CreateArticleStep1} />
-            <Route path="/createArticle/step2" component={CreateArticleStep2} />
+            {/* 이건 사용될 일 없음 (나중에 제거해도 됨) */}
+            <Route path="/createArticle/step2" exact />
+            <Route path="/createArticle/step2/:type" component={CreateArticleStep2} />
             <Route path="/createArticle/step3" component={CreateArticleStep3} />
         </Switch>
     );
