+++ agaein_web/src/router/params.ts
@@ -1,5 +1,10 @@
-export interface CreateArticleParams {
-    id: string;
+import { Board_Type } from 'graphql/generated/generated';
+
+export interface CreateArticleStep1Params {
+    id?: string;
+}
+export interface CreateArticleStep2Params {
+    type: Board_Type;
 }
 export interface ArticleListParams {
     id: string;
